renderercommon: fix warning

+++ src/renderercommon/tr_public.h
@@ -182,7 +182,7 @@ typedef struct
 #else
 	void * (*Hunk_Alloc)(size_t size, ha_pref pref);
 #endif
-	void * (*Hunk_AllocateTempMemory)(size_t size);
+	void * (*Hunk_AllocateTempMemory)(unsigned int size);
 	void (*Hunk_FreeTempMemory)(void *block);
 
 	// dynamic memory allocator for things that need to be freed
